requirements.txt
aiofiles==23.1.0
asyncpg==0.27.0
alembic==1.9.2
fastapi~=0.89.1
Jinja2~=3.1.2
pre-commit==3.0.4
pydantic~=1.10.4
python-dotenv==0.21.1
python-multipart~=0.0.5
starlette~=0.22.0
SQLAlchemy==1.4.46
uvicorn==0.20.0
python-jose[cryptography]
passlib[bcrypt]
